# Concurrent Reader::read_tiles()