@joe-maley 2020-09-18
I tested the same scenario on an EC2 instance, the bottleneck is still the read path:

```
==== READ ====

- Number of read queries: 1
- Number of attempts until results are found: 1

- Number of attributes read: 15
  * Number of fixed-sized attributes read: 7
  * Number of var-sized attributes read: 8
- Number of dimensions read: 2
  * Number of fixed-sized dimensions read: 2

- Number of logical tiles overlapping the query: 49
- Number of physical tiles read: 833
  * Number of physical fixed-sized tiles read: 441
  * Number of physical var-sized tiles read: 784
- Number of cells read: 490000
- Number of result cells: 342491
- Percentage of useful cells read: 69.8961%

- Number of bytes read: 12203857 bytes (0.0113657 GB)
- Number of read operations: 1795
- Number of bytes unfiltered: 85702177 bytes (0.0798164 GB)
- Unfiltering inflation factor: 7.02255x

- Time to compute estimated result size: 0.168786 secs
  * Time to compute tile overlap: 0.131002 secs
    > Time to compute relevant fragments: 3.5985e-05 secs
    > Time to load relevant fragment R-trees: 0.130686 secs
    > Time to compute relevant fragment tile overlap: 0.000266461 secs

- Time to open array: 0.265584 secs
  * Time to load array schema: 0.0425474 secs
  * Time to load consolidated fragment metadata: 1.144e-06 secs
  * Time to load fragment metadata: 0.1728 secs

- Total metadata read: 3435806 bytes (0.00319984 GB)
  * Array schema: 674 bytes (6.27711e-07 GB)
  * Fragment metadata: 10948 bytes (1.01961e-05 GB)
  * R-tree: 406928 bytes (0.000378981 GB)
  * Fixed-sized tile offsets: 1554344 bytes (0.0014476 GB)
  * Var-sized tile offsets: 731456 bytes (0.000681221 GB)
  * Var-sized tile sizes: 731456 bytes (0.000681221 GB)

- Time to initialize the read state: 3.9915e-05 secs

- Read time: 2.64719 secs
  * Time to compute next partition: 0.000460646 secs
  * Time to compute result coordinates: 0.312875 secs
    > Time to compute sparse result tiles: 0.000109849 secs
    > Time to read coordinate tiles: 0.286124 secs
    > Time to unfilter coordinate tiles: 0.00523176 secs
    > Time to compute range result coordinates: 0.0145431 secs
  * Time to compute sparse result cell slabs: 0.00858648 secs
  * Time to copy result attribute values: 2.31396 secs
    > Time to read attribute tiles: 3.13746 secs
    > Time to unfilter attribute tiles: 0.0512189 secs
    > Time to copy fixed-sized attribute values: 0.0410954 secs
    > Time to copy var-sized attribute values: 0.269066 secs
  * Time to copy result coordinates: 0.0106431 secs
    > Time to copy fixed-sized coordinates: 0.010537 secs

- Total read query time (array open + init state + read): 2.64723 secs


==== Python ====
- Read query time: 2.81927
- Buffer conversion time: 0.0830522
```