@joe-maley 2020-09-18
@stavrospapadopoulos I modified the stat timer to start outside of the concurrent loop (see the new `read_attribute_tiles` and `read_coordinate_tiles`. Here's an example:
```
  * Time to copy result attribute values: 3.31069 secs
    > Time to read attribute tiles: 2.92615 secs
```